resize doctype

+++ webEdition/liveUpdate/sqldumps/tblnavigationrules.sql
@@ -4,7 +4,7 @@ CREATE TABLE ###TBLPREFIX###tblnavigationrules (
   NavigationID int unsigned NOT NULL default '0',
   SelectionType varchar(16) NOT NULL default '',
   FolderID int unsigned NOT NULL default '0',
-  DoctypeID int unsigned NOT NULL default '0',
+  DoctypeID mediumint NOT NULL default '0',
   Categories text NOT NULL,
   ClassID int unsigned NOT NULL default '0',
   WorkspaceID int unsigned NOT NULL default '0',
